useFagsakApi hadde en kobling til FagsakContext, som igjen hadde en kobling til useFagsakApi, bryter sirkulær dependency
